chore: type 위치 변경
